context

+++ rengine-src/Rengine/ImGui/ImGuiLayer.cpp
@@ -1,10 +1,11 @@
+#include "repch.hpp"
 #include "ImGuiLayer.hpp"
 #include "imgui.h"
 #include "backends/imgui_impl_glfw.h"
 #include "backends/imgui_impl_opengl3.h"
 #include "backends/imgui_impl_opengl3_loader.h"
 
-#include "../application.hpp"
+#include "Rengine/application.hpp"
 
 // #include <glad/glad.h>
 #include <GLFW/glfw3.h>
@@ -68,7 +69,7 @@ void ImGuiLayer::End()
 
     auto& app = Application::getApplication();
 
-    io.DisplaySize = ImVec2(app.getWindow().getWidth(),app.getWindow().getHeight());
+    io.DisplaySize = ImVec2(static_cast<float>(app.getWindow().getWidth()),static_cast<float>(app.getWindow().getHeight()));
 
     ImGui::Render();
